+++ Electronics Toolbox/ContentView.swift
@@ -8,6 +8,7 @@
 import SwiftUI
 
 struct ContentView: View {
+    
     var body: some View {
         CalculationList()
     }
@@ -17,4 +18,19 @@ struct ContentView_Previews: PreviewProvider {
     static var previews: some View {
         ContentView()
     }
+}
+
+func successHaptics() {
+    let generator = UINotificationFeedbackGenerator()
+    generator.notificationOccurred(.success)
+}
+
+func errorHaptics() {
+    let generator = UINotificationFeedbackGenerator()
+    generator.notificationOccurred(.error)
+}
+
+func selectionHaptics() {
+    let generator = UISelectionFeedbackGenerator()
+    generator.selectionChanged()
 }